[refactor] position,span,asciifile to own files; remove PositionedChar
This is prelimenary work for save handling of text ranges and indices. This commit
enforces a separation between asciifile::* (which is now the only one allowed to access indices)
and the rest of the code.

- Since Position carries a reference to asciifile, PositionedChar is just a reshaped
  version of Position (with maybe better cache behaviour). We remove PositionedChar
  internally, but keep the Iterator and the type definition to not break existing code.
  (which will be changed in a later refator commit.)
- We make inherently unsafe APIs that expose positions in the form of indices private.
  As a result Lexer fails to compile, because it is constructing spans from indices.
  (This is a workaround for peek() not returning Position or Span along the data. This
  will be changed in an upcoming commit.)
- Position started out as a simple data object containing `col:usize` and `row:usize`.
  It has been carrying a reference to the file in an additional field for a while now.
  This means we can actually implement a rich API on it, e.g. exposing the character
  at the position, a linked list/iterator to the next/previous characters, a check
  for EOF, etc.
